requirements.txt
pytz==2016.4
django==1.9.6
python-dateutil==2.5.3
django-taggit==0.20.2
djangorestframework==3.6.3
django-taggit-serializer==0.1.5
